+++ sdk/include/opentelemetry/sdk/metrics/state/attributes_hashmap.h
@@ -49,7 +49,13 @@ class AttributesHashMapWithCustomHash
 public:
   AttributesHashMapWithCustomHash(size_t attributes_limit = kAggregationCardinalityLimit)
       : attributes_limit_(attributes_limit)
-  {}
+  {
+    if (attributes_limit_ > kAggregationCardinalityLimit)
+    {
+      hash_map_.reserve(attributes_limit_);
+    }
+  }
+
   Aggregation *Get(const MetricAttributes &attributes) const
   {
     auto it = hash_map_.find(attributes);