+++ CHANGELOG.md
@@ -33,6 +33,10 @@ Increment the:
 * [BUILD] Use -dev versions in main branch
   [#3609](https://github.com/open-telemetry/opentelemetry-cpp/pull/3609)
 
+* [SDK] Implementing configurable aggregation cardinality limit and collector
+  emitting all views instead of last view
+  [#3624](https://github.com/open-telemetry/opentelemetry-cpp/pull/3624)
+
 Important changes:
 
 * [CMAKE] Upgrade CMake minimum version to 3.16